refactor: TicketResultActivity View 세팅 코드 분리
