+++ app/src/main/java/woowacourse/movie/activity/ticketresult/TicketResultActivity.kt
@@ -2,16 +2,11 @@ package woowacourse.movie.activity.ticketresult
 
 import android.os.Bundle
 import android.view.MenuItem
-import android.widget.TextView
 import android.widget.Toast
 import androidx.appcompat.app.AppCompatActivity
 import woowacourse.movie.R
 import woowacourse.movie.model.TicketModel
 import woowacourse.movie.util.getSerializableExtraCompat
-import java.text.DecimalFormat
-import java.time.LocalDate
-import java.time.LocalTime
-import java.time.format.DateTimeFormatter
 
 class TicketResultActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
@@ -27,37 +22,7 @@ class TicketResultActivity : AppCompatActivity() {
             finish()
             return
         }
-        initTitle(ticket.title)
-        initPlayingDate(ticket.playingDate, ticket.playingTime)
-        initCount(ticket.count)
-        initPricePayment(ticket.price)
-    }
-
-    private fun initTitle(title: String) {
-        val titleView = findViewById<TextView>(R.id.text_title)
-        titleView.text = title
-    }
-
-    private fun initPlayingDate(playingDate: LocalDate, playingTime: LocalTime) {
-        val playingDateView = findViewById<TextView>(R.id.text_playing_date)
-        playingDateView.text = getString(
-            R.string.date_time,
-            DateTimeFormatter.ofPattern(getString(R.string.date_format)).format(playingDate),
-            DateTimeFormatter.ofPattern(getString(R.string.time_format)).format(playingTime)
-        )
-    }
-
-    private fun initCount(count: Int) {
-        val countView = findViewById<TextView>(R.id.text_person_count)
-        countView.text = getString(R.string.normal_count, count)
-    }
-
-    private fun initPricePayment(price: Int) {
-        val pricePayment = findViewById<TextView>(R.id.text_price_payment)
-        pricePayment.text = getString(
-            R.string.price_payment,
-            DecimalFormat(getString(R.string.decimal_format)).format(price)
-        )
+        TicketView(findViewById(R.id.layout_ticket)).set(ticket)
     }
 
     override fun onOptionsItemSelected(item: MenuItem): Boolean {
