+++ app/src/main/res/layout/activity_ticket_result.xml
@@ -5,7 +5,8 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".activity.ticketresult.TicketResultActivity">
+    tools:context=".activity.ticketresult.TicketResultActivity"
+    android:id="@+id/layout_ticket">
 
     <TextView
         android:id="@+id/text_notice"
